chore: move wallet client hooks out of experimental
