--- account-kit/core/src/experimental/README.md
@@ -1,3 +1,0 @@
-# Experimental
-
-This experimental namespace is meant to contain actions and other core logic that consumes Wallet APIs. These will replace the usual flow found within this repo
