--- account-kit/core/src/experimental/index.ts
@@ -1,4 +1,0 @@
-export type * from "./actions/getSmartWalletClient.js";
-export { getSmartWalletClient } from "./actions/getSmartWalletClient.js";
-export type * from "./actions/watchSmartWalletClient.js";
-export { watchSmartWalletClient } from "./actions/watchSmartWalletClient.js";
