+++ account-kit/react/src/hooks/useSignMessage.ts
@@ -11,7 +11,7 @@ import { useAlchemyAccountContext } from "./useAlchemyAccountContext.js";
 import { ClientUndefinedHookError } from "../errors.js";
 import { ReactLogger } from "../metrics.js";
 import type { BaseHookMutationArgs } from "../types.js";
-import { useSmartWalletClient } from "../experimental/hooks/useSmartWalletClient.js";
+import { useSmartWalletClient } from "./useSmartWalletClient.js";
 import { useAccount as wagmi_useAccount } from "wagmi";
 import { signMessage as wagmi_signMessage } from "@wagmi/core";
