+++ account-kit/react/src/hooks/useWaitForUserOperationTransaction.ts
@@ -12,7 +12,7 @@ import { useAlchemyAccountContext } from "./useAlchemyAccountContext.js";
 import { ClientUndefinedHookError } from "../errors.js";
 import { ReactLogger } from "../metrics.js";
 import type { BaseHookMutationArgs } from "../types.js";
-import { useSmartWalletClient } from "../experimental/hooks/useSmartWalletClient.js";
+import { useSmartWalletClient } from "./useSmartWalletClient.js";
 import type { UseSmartAccountClientResult } from "./useSmartAccountClient.js";
 
 export type UseWaitForUserOperationTransactionMutationArgs =
