+++ account-kit/react/src/index.ts
@@ -67,6 +67,20 @@ export type * from "./hooks/useUser.js";
 export { useUser } from "./hooks/useUser.js";
 export type * from "./hooks/useWaitForUserOperationTransaction.js";
 export { useWaitForUserOperationTransaction } from "./hooks/useWaitForUserOperationTransaction.js";
+export { useCallsStatus } from "./hooks/useCallsStatus.js";
+export type * from "./hooks/useCallsStatus.js";
+export { useGrantPermissions } from "./hooks/useGrantPermissions.js";
+export type * from "./hooks/useGrantPermissions.js";
+export { usePrepareCalls } from "./hooks/usePrepareCalls.js";
+export type * from "./hooks/usePrepareCalls.js";
+export { useSendCalls } from "./hooks/useSendCalls.js";
+export type * from "./hooks/useSendCalls.js";
+export { useSendPreparedCalls } from "./hooks/useSendPreparedCalls.js";
+export type * from "./hooks/useSendPreparedCalls.js";
+export { useSmartWalletClient } from "./hooks/useSmartWalletClient.js";
+export type * from "./hooks/useSmartWalletClient.js";
+export { useWaitForCallsStatus } from "./hooks/useWaitForCallsStatus.js";
+export type * from "./hooks/useWaitForCallsStatus.js";
 export { type AlchemyAccountsUIConfig } from "./types.js";
 export { Dialog } from "./components/dialog/dialog.js";
